refactor: Extract StackRowView to separate file
Move StackRowView from HomeView.swift to its own file to comply with
SwiftLint file_length rule (max 500 lines).
